initial commit
this commit reflects a transfer of the chatgpt draft into posts. the draft was copied and modified slightly so that the images could be viewable.

in other words, the images are now located in assets/2022-12-01-chatgpt and the '_posts' version of the chatgpt article was modified so that the paths to images were compatible

a source of possible issue, @mcnewcp used a  convetion in front of the 'assets' part of the file path. @907Resident is unsure of what that means but still expects the images to show up properly.  is a directory hidden and only @mcnewcp can access it